Partial solution to re-enabling OpenSciView button
Pressing sync checks and reenables the button. Otherwise, I think
we need to have a listener for the SciView windowclosed event

+++ src/main/java/sc/fiji/snt/SciViewSNT.java
@@ -118,33 +118,37 @@ public class SciViewSNT {
 		if (SNTUtils.getPluginInstance().getPathAndFillManager().size() == 0) {
 			return false;
 		}
-		final Tree tree = new Tree(SNTUtils.getPluginInstance().getPathAndFillManager()
-				.getPathsFiltered());
-		if (plottedTrees.containsKey(PATH_MANAGER_TREE_LABEL)) {// PATH_MANAGER_TREE_LABEL, the value of this is the *new* tree to add
-			// TODO If the Node exists, then remove and add new one to replace
-			//System.out.println("Tree exists, updating current name: " + tree.getLabel() + " next label: " + PATH_MANAGER_TREE_LABEL);
-			//System.out.println("Deleting nodes: " + plottedTrees.get(PATH_MANAGER_TREE_LABEL).getChildren().size());
-			for( Node node : plottedTrees.get(PATH_MANAGER_TREE_LABEL).getChildren() ) {
-				//syncNode(tree,node);
-				plottedTrees.get(PATH_MANAGER_TREE_LABEL).removeChild(node);
-				sciView.deleteNode(node, false);
+		if( sciView.isClosed() ) {
+			SNTUtils.getPluginInstance().getUI().setOpenSciViewButtonEnabled(true);
+		} else {
+
+			final Tree tree = new Tree(SNTUtils.getPluginInstance().getPathAndFillManager()
+					.getPathsFiltered());
+			if (plottedTrees.containsKey(PATH_MANAGER_TREE_LABEL)) {// PATH_MANAGER_TREE_LABEL, the value of this is the *new* tree to add
+				// TODO If the Node exists, then remove and add new one to replace
+				//System.out.println("Tree exists, updating current name: " + tree.getLabel() + " next label: " + PATH_MANAGER_TREE_LABEL);
+				//System.out.println("Deleting nodes: " + plottedTrees.get(PATH_MANAGER_TREE_LABEL).getChildren().size());
+				for (Node node : plottedTrees.get(PATH_MANAGER_TREE_LABEL).getChildren()) {
+					//syncNode(tree,node);
+					plottedTrees.get(PATH_MANAGER_TREE_LABEL).removeChild(node);
+					sciView.deleteNode(node, false);
+				}
+				sciView.deleteNode(plottedTrees.get(PATH_MANAGER_TREE_LABEL));
+				plottedTrees.remove(PATH_MANAGER_TREE_LABEL);
+				//System.out.println("Num remaining nodes: " + sciView.getSceneNodes().length);
+
+				// Dont create a new SNT node each time
+
+				tree.setLabel(PATH_MANAGER_TREE_LABEL);
+				add(tree);
+
+				//System.out.println("Num remaining nodes: " + sciView.getSceneNodes().length);
+				//sciView.centerOnNode( plottedTrees.get(PATH_MANAGER_TREE_LABEL) );
+				//sciView.centerOnNode( sciView.getSceneNodes()[(int)(Math.random()*sciView.getSceneNodes().length)] );
+			} else {
+				tree.setLabel(PATH_MANAGER_TREE_LABEL);
+				add(tree);
 			}
-			sciView.deleteNode(plottedTrees.get(PATH_MANAGER_TREE_LABEL));
-			plottedTrees.remove(PATH_MANAGER_TREE_LABEL);
-			//System.out.println("Num remaining nodes: " + sciView.getSceneNodes().length);
-
-			// Dont create a new SNT node each time
-
-			tree.setLabel(PATH_MANAGER_TREE_LABEL);
-			add(tree);
-
-			//System.out.println("Num remaining nodes: " + sciView.getSceneNodes().length);
-			//sciView.centerOnNode( plottedTrees.get(PATH_MANAGER_TREE_LABEL) );
-			//sciView.centerOnNode( sciView.getSceneNodes()[(int)(Math.random()*sciView.getSceneNodes().length)] );
-		}
-		else {
-			tree.setLabel(PATH_MANAGER_TREE_LABEL);
-			add(tree);
 		}
 		return true;
 	}
